Restructure the whole NetworkManager code to be single component & updated COMRPC Proxy Implementation
