+++ NetworkManager/INetworkManager.h
