+++ NetworkManager/LegacyNetwork.h
