+++ NetworkManager/LegacyWiFi.h
