+++ NetworkManager/Module.cpp
