+++ NetworkManager/NetworkManager.config
@@ -5,7 +5,6 @@ map()
    key(root)
    map()
       kv(outofprocess true)
-      kv(locator lib${PLUGIN_IMPLEMENTATION}.so)
    end()
    key(connectivity)
    map()
