+++ NetworkManager/NetworkManager.json
