+++ NetworkManager/NetworkManagerDeprecatedAPIs.conf.in
