+++ NetworkManager/NetworkManagerPlugin.json
