+++ NetworkManager/WiFiManagerDeprecatedAPIs.conf.in
