+++ NetworkManager/WiFiManagerDeprecatedAPIs.config
