+++ NetworkManager/WiFiManagerDeprecatedAPIs.cpp
