+++ NetworkManager/WiFiSignalStrengthMonitor.h
@@ -32,12 +32,12 @@ namespace WPEFramework
 {
     namespace Plugin
     {
-        class WifiSignalStrengthMonitor
+        class WiFiSignalStrengthMonitor
         {
             public:
-                WifiSignalStrengthMonitor():isRunning(false) {}
-                ~WifiSignalStrengthMonitor(){ NMLOG_INFO("~WifiSignalStrengthMonitor"); }
-                void startWifiSignalStrengthMonitor(int interval);
+                WiFiSignalStrengthMonitor():isRunning(false) {}
+                ~WiFiSignalStrengthMonitor(){ NMLOG_INFO("~WiFiSignalStrengthMonitor"); }
+                void startWiFiSignalStrengthMonitor(int interval);
                 void getSignalData(std::string &ssid, Exchange::INetworkManager::WiFiSignalQuality &quality, std::string &strengthOut);
             private:
                 std::string retrieveValues(const char *command, const char* keyword, char *output_buffer, size_t output_buffer_size);
